pyproject.toml
[tool.poetry]
name = "python-api-tests"
version = "0.1.0"
description = "Python Api tests"
authors = ["Alexander Lozovoy <berpress@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pre-commit = "^2.15.0"
flake8 = "^3.9.2"
pytest = "^6.2.5"
requests = "^2.26.0"
Faker = "^8.12.1"
cattrs = "^1.8.0"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
